distribute the branches
