+++ src/mvs/depth_map.cc
@@ -1,0 +1,4 @@
+#include "depthmap.h"
+
+float DepthMap::getDepthMin() const { return depth_min_; }
+float DepthMap::getDepthMax() const { return depth_max_; }
